LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-navigated-history-popup.https-expected.txt


FAIL Popup navigating to other origin with COOP: same-origin and back in history assert_equals: name expected "" but got "Popup-navigating-to-other-origin-with-COOP:-same-origin-and-back-in-history"
